Update para versão 1.5

+++ dspace-setup/dspace-webservice.sh
@@ -3,11 +3,12 @@
 # =================================================
 # MÓDULO DE CONFIGURAÇÃO DA APLICAÇÃO WEB
 # =================================================
-# dspace-webservice, version 1.0, UFPA | 2017
+# dspace-webservice, version 1.1, UFPA | 2017
 
 # Argumentos de entrada:
 #	$1 - diretório base do DSpace
 #	$2 - diretório 'home/$user' do usuário padrão do DSpace
+#	$3 - caminho da logo 
 
 # Atribuições locais
 base_dir=$1
@@ -23,9 +24,12 @@ writewarning $? "Não foi possível copiar os arquivos de tradução"
 
 # Configura a logo da página inicial do repositório
 if [ "$logo" != "" ]; then
-	cp "$logo" "$home/logo.png"
-	mv "$home/logo.png" /$base_dir/webapps/jspui/image/ -f
-	writewarning $? "Problema ao configurar logotipo em 'webapps'"
+	if [ -e $logo ]; then
+		sed -i "s/<img src=\"<%= request.getContextPath() %>.*/<img src=\"<%= request.getContextPath() %>\/image\/logo.png\" usemap=\"#mapa-brasil\" style=\"max-height:300px; cursor:pointer;margin:20px 0px;\" class=\"image-logo\">/" /$base_dir/webapps/jspui/home.jsp
+		cp "$logo" "$home/logo.png"
+		mv "$home/logo.png" /$base_dir/webapps/jspui/image/ -f
+		writewarning $? "Problema ao configurar logotipo em 'webapps'"
+	fi
 fi
 
 # Configurações da aplicação web no Apache Tomcat
